+++ c_json_parser/tests/unit_tests/parse_object_unit_test.c
@@ -4,7 +4,7 @@
 #include "unit_tests.h"
 #define N 6
 
-void make_object_test(int expected, char input[])
+void make_object_test(size_t expected, char input[])
 {
     JSONRoot root;
     if( parse_json(&root, input) == -1 )
@@ -33,7 +33,7 @@ void make_object_test(int expected, char input[])
 void unit_test_parse_object()
 {
     char input[N][100] = { "{  }", "{\"something\":3}", "{\"3\":32,\"res\":23}", "{ \"44wfc\":[],\"r1vrve\":[3,2,1] }", "{\"res\": {}}", "{\"fe\":\"fr\", \"ce\": \"fr\"}"};
-    int expected[N] = {0, 1, 2, 2, 1, 2};
+    size_t expected[N] = {0, 1, 2, 2, 1, 2};
 
     for( int i = 0; i < N; i++ )
         make_object_test(expected[i], input[i]);
